- Исправления 1.16.4 - Причесал весь проект по MVVM.
