+++ Tracker/ViewControllers/TrackerVC/Views/TrackerVC+HideDatePicker.swift
@@ -1,0 +1,32 @@
+//
+//  TrackerVC+HideDatePicker.swift
+//  Tracker
+//
+//  Created by Kirill Sklyarov on 09.04.2024.
+//
+
+import UIKit
+
+// MARK: - HideDatePicker
+extension TrackerViewController: UIGestureRecognizerDelegate {
+    
+    func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer, shouldReceive touch: UITouch) -> Bool {
+        let location = touch.location(in: view)
+        return !datePicker.frame.contains(location)
+    }
+    
+    func addTapGestureToHideDatePicker() {
+        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(handleTap))
+        tapGesture.delegate = self
+        view.addGestureRecognizer(tapGesture)
+    }
+    
+    @objc func handleTap(_ sender: UITapGestureRecognizer) {
+        let location = sender.location(in: view)
+        
+        if !datePicker.frame.contains(location) {
+            datePicker.isHidden = true
+        }
+    }
+}
+
